Fix filename from uri
