+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/composable/ComposableReceive.kt
@@ -10,6 +10,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import com.flipperdevices.deeplink.model.DeeplinkContent
 import com.flipperdevices.share.common.composable.ComposableAlertDialog
 import com.flipperdevices.share.receive.R
+import com.flipperdevices.share.receive.util.filename
 import com.flipperdevices.share.receive.viewmodel.ReceiveViewModel
 import com.flipperdevices.share.receive.viewmodel.ReceiveViewModelFactory
 
@@ -39,7 +40,7 @@ fun ComposableReceive(
     ComposableAlertDialog(
         title = stringResource(
             R.string.receive_dialog_title,
-            deeplinkContent.filename() ?: ""
+            deeplinkContent.filename(LocalContext.current.contentResolver) ?: ""
         ),
         downloadProgress = receiveState.downloadProgress,
         onCancel = {
